move symbol handling to own file
