+++ src/main.cpp
@@ -23,6 +23,7 @@
 #include "messagebuffer.hpp"
 #include "workspace.hpp"
 #include "utils.hpp"
+#include "symbols.hpp"
 
 using json = nlohmann::json;
 namespace fs = std::experimental::filesystem;
@@ -158,334 +159,6 @@ json get_diagnostics(std::string uri, std::string content,
     return diagnostics;
 }
 
-/// Returns the byte offset for the given character on the given line.
-// FIXME: use UTF-16 offsets
-// https://fasterthanli.me/articles/the-bottom-emoji-breaks-rust-analyzer
-int find_position_offset(const char* text, int line, int character) {
-    int offset = 0;
-    while (line > 0) {
-        while (text[offset] && text[offset] != '\n') offset += 1;
-        offset += text[offset] == '\n';
-        line -= 1;
-    }
-
-    while (character > 0 && text[offset] && text[offset] != '\n') {
-        offset += 1;
-        character -= 1;
-    }
-
-    return offset;
-}
-
-/// Returns `true` if the character may start an identifier.
-bool is_identifier_start_char(char c) {
-    return ('a' <= c && c <= 'z') || ('A' <= c && c <= 'Z') || c == '_';
-}
-
-/// Returns `true` if the character may be part of an identifier.
-bool is_identifier_char(char c) {
-    return is_identifier_start_char(c) || ('0' <= c && c <= '9');
-}
-
-/// Returns the offset in `text` where the last word started.
-int get_last_word_start(const char* text, int offset) {
-    int start = offset;
-    while (start > 0 && is_identifier_char(text[start - 1])) {
-        start -= 1;
-    }
-
-    // If `text` was `123abc` and `offset` pointed at `b`, start would point at `1`.
-    // We want to point to `a`, so advance past any characters that are not a
-    // valid start of an identifier.
-    while (start < offset && !is_identifier_start_char(text[start])) {
-        start += 1;
-    }
-
-    return start;
-}
-
-int get_word_end(const char* text, int start) {
-    int end = start;
-    while (text[end] && is_identifier_char(text[end])) end++;
-    return end;
-}
-
-struct Symbol {
-    enum Kind {
-        Unknown = 0,
-        Function = 3,
-        Type = 7,
-        Constant = 21,
-    };
-
-    Kind kind = Unknown;
-    std::string details;
-};
-
-typedef std::map<std::string, Symbol> SymbolMap;
-
-void add_builtin_types(SymbolMap& symbols)  {
-    symbols.emplace("bool", Symbol{Symbol::Type, "<type>"});
-    symbols.emplace("int", Symbol{Symbol::Type, "<type>"});
-    symbols.emplace("uint", Symbol{Symbol::Type, "<type>"});
-    symbols.emplace("float", Symbol{Symbol::Type, "<type>"});
-    symbols.emplace("double", Symbol{Symbol::Type, "<type>"});
-
-    std::string vec_buffer = "TvecX";
-    for (int i = 2; i <= 4; i++) {
-        vec_buffer[4] = '0' + i;
-
-        // vec2, vec3, vec4, etc.
-        symbols.emplace(&vec_buffer[1], Symbol{Symbol::Type, "<type>"});
-        // bvecX, ivecX, uvecX, dvecX
-        vec_buffer[0] = 'b';
-        symbols.emplace(vec_buffer, Symbol{Symbol::Type, "<type>"});
-        vec_buffer[0] = 'i';
-        symbols.emplace(vec_buffer, Symbol{Symbol::Type, "<type>"});
-        vec_buffer[0] = 'u';
-        symbols.emplace(vec_buffer, Symbol{Symbol::Type, "<type>"});
-        vec_buffer[0] = 'd';
-        symbols.emplace(vec_buffer, Symbol{Symbol::Type, "<type>"});
-    }
-
-    std::string mat_buffer = "dmatXxX";
-    for (int col = 2; col <= 4; col++) {
-        mat_buffer[4] = '0' + col;
-
-        for (int row = 2; row <= 4; row++) {
-            mat_buffer[6] = '0' + row;
-            symbols.emplace(&mat_buffer[1], Symbol{Symbol::Type, "<type>"});
-            symbols.emplace(mat_buffer, Symbol{Symbol::Type, "<type>"});
-        }
-
-        mat_buffer[5] = 0;
-        symbols.emplace(&mat_buffer[0], Symbol{Symbol::Type, "<type>"});
-        symbols.emplace(&mat_buffer[1], Symbol{Symbol::Type, "<type>"});
-        mat_buffer[5] = 'x';
-    }
-    
-    const char* image_kinds[] = {
-        "1D",
-        "2D",
-        "3D",
-        "Cube",
-        "2DRect",
-        "1DArray",
-        "2DArray",
-        "CubeArray",
-        "Buffer",
-        "2DMS",
-        "2DMSArray",
-    };
-
-    int image_count = sizeof(image_kinds) / sizeof(image_kinds[0]);
-    for (int i = 0; i < image_count; i++) {
-        std::string buffer = "gimage";
-        buffer += image_kinds[i];
-
-        symbols.emplace(&buffer[1], Symbol{Symbol::Type, "<type>"});
-        buffer[0] = 'i';
-        symbols.emplace(buffer, Symbol{Symbol::Type, "<type>"});
-        buffer[0] = 'u';
-        symbols.emplace(buffer, Symbol{Symbol::Type, "<type>"});
-    }
-
-    for (int i = 0; i < image_count; i++) {
-        std::string buffer = "gsampler";
-        buffer += image_kinds[i];
-
-        symbols.emplace(&buffer[1], Symbol{Symbol::Type, "<type>"});
-        buffer[0] = 'i';
-        symbols.emplace(buffer, Symbol{Symbol::Type, "<type>"});
-        buffer[0] = 'u';
-        symbols.emplace(buffer, Symbol{Symbol::Type, "<type>"});
-    }
-
-    const char* shadow_samplers[] = {
-        "sampler1DShadow",
-        "sampler2DShadow",
-        "samplerCubeShadow",
-        "sampler2DRectShadow",
-        "sampler1DArrayShadow",
-        "sampler2DArrayShadow",
-        "samplerCubeArrayShadow",
-    };
-    int shadow_sampler_count = sizeof(shadow_samplers) / sizeof(shadow_samplers[0]);
-    for (int i = 0; i < shadow_sampler_count; i++) {
-        symbols.emplace(shadow_samplers[i], Symbol{Symbol::Type, "<type>"});
-    }
-}
-
-struct Word {
-    const char* start = nullptr;
-    const char* end = nullptr;
-
-    bool is_equal(const char* text) const {
-        const char* s = start;
-        while (s != end && *s == *text) {
-            s++;
-            text++;
-        }
-        return s == end && *text == 0;
-    }
-};
-
-bool is_whitespace(char c) {
-    return c == ' ' || c == '\t' || c == '\n';
-}
-
-/// Extracts all symbols from the given string, and inserts them into the symbol map.
-///
-/// The current implementation is very naive and may not handle certain cases that well.
-void extract_symbols(const char* text, SymbolMap& symbols) {
-    std::vector<Word> words;
-    int arguments = 0;
-    bool had_arguments = false;
-    Word inside_block{};
-
-    const char* p = text;
-    while (*p) {
-        if (is_identifier_start_char(*p)) {
-            const char* start = p;
-            while (is_identifier_char(*p)) p++;
-            if (*p == '[') {
-                while (*p && *p != ']') p++;
-            }
-            Word ident{start, p};
-
-            // don't confuse `layout(...)` for a function.
-            if (ident.is_equal("layout")) {
-                while (is_whitespace(*p)) p++;
-                if (*p == '(') {
-                    while (*p && *p != ')') p++;
-                }
-                continue;
-            }
-
-            words.push_back(ident);
-            continue;
-        } 
-
-        // don't confuse numeric literals as identifiers
-        if ('0' <= *p && *p <= '9') {
-            p++;
-            while (is_identifier_char(*p)) p++;
-            continue;
-        } 
-
-        if (*p == '{') {
-            // TODO: handle function bodies
-
-            if (words.size() >= 2 && arguments == 0) {
-                Word kind = words[words.size() - 2];
-                if (kind.is_equal("in") 
-                        || kind.is_equal("out") 
-                        || kind.is_equal("uniform") 
-                        || kind.is_equal("buffer")) {
-                    inside_block = words[words.size() - 1];
-                    words.clear();
-                    p++;
-                    continue;
-                }
-            }
-
-            // skip struct fields and function bodies (their contents are not global)
-            while (*p && *p != '}') p++;
-            continue;
-        } 
-
-        if (*p == '}' && inside_block.start) {
-            words.push_back(inside_block);
-            inside_block = Word{};
-        }
-
-        if (*p == '(') {
-            had_arguments = true;
-            p++;
-            const char* start = nullptr;
-            const char* end = nullptr;
-            while (*p) {
-                if (is_whitespace(*p)) {
-                    p++;
-                    continue;
-                }
-
-                if (*p == ')' || *p == ',') {
-                    if (start) {
-                        words.push_back({start, p});
-                        arguments++;
-                    }
-
-                    if (*p == ')') break;
-
-                    p++;
-                    start = nullptr;
-                    end = nullptr;
-                    continue;
-                }
-
-                if (!start) start = p;
-                end = p;
-
-                p++;
-            }
-        } 
-
-        if (*p == ';' || *p == ')') {
-            // end of declaration
-            int name_index = (int)words.size() - arguments - 1;
-            int type_index = name_index - 1;
-
-            if (name_index >= 0) {
-                Word name_word = words[name_index];
-                Word type_word = type_index >= 0 ? words[type_index] : Word{};
-
-                std::string name(name_word.start, name_word.end);
-                std::string type(type_word.start, type_word.end);
-
-                if (!type.empty()) {
-                    symbols.emplace(type, Symbol{Symbol::Type, "<type>"});
-                }
-
-                for (int i = 0; i < arguments; i++) {
-                    if (i == 0) {
-                        type += " (";
-                    } else {
-                        type += ", ";
-                    }
-
-
-                    Word arg = words[name_index + 1 + i];
-                    const char* t = arg.start;
-                    while (t != arg.end) {
-                        if (is_whitespace(*t)) {
-                            // only emit a single space
-                            type.push_back(' ');
-                            while (t != arg.end && is_whitespace(*t)) t++;
-                        } else {
-                            type.push_back(*t);
-                            t++;
-                        }
-                    }
-
-                    if (i == arguments - 1) {
-                        type += ")";
-                    }
-                }
-
-                Symbol::Kind kind = *p == ')' ? Symbol::Function : Symbol::Constant;
-                symbols.emplace(name, Symbol{kind, type});
-            }
-
-            words.clear();
-            arguments = 0;
-            had_arguments = false;
-        }
-
-        p++;
-    }
-}
-
 SymbolMap get_symbols(const std::string& uri, AppState& appstate){
     auto language = find_language(uri);
 
@@ -522,9 +195,6 @@ SymbolMap get_symbols(const std::string& uri, AppState& appstate){
 void find_completions(const SymbolMap& symbols, const std::string& prefix, std::vector<json>& out) {
     for (auto& entry : symbols) {
         auto& name = entry.first;
-        if (name.size() < prefix.size()) continue;
-        if (strncmp(name.c_str(), prefix.c_str(), prefix.size()) != 0) continue;
-
         auto& symbol = entry.second;
         out.push_back(json {
             { "label", name },
